tests/baseline/protocol-1.2/inspector/console/console-command-clear-expected.txt

CONSOLE MESSAGE: line 9: one
CONSOLE MESSAGE: line 10: two
CONSOLE MESSAGE: line 11: three
CONSOLE MESSAGE: line 1: console.clear
Tests that console is cleared upon clear() eval in console.

=== Before clear ===
console-command-clear.html:9 one
console-command-clear.html:10 two
console-command-clear.html:11 three
=== After clear ===
VM:1 Console was cleared
